Changed the data type from int to float for one of the lists

+++ intersect.py
@@ -1,4 +1,4 @@
-group_One= [float(a) for a in input("Enter group number 1:"). split()]
+group_One= [int(a) for a in input("Enter group number 1:"). split()]
 group_Two= [float(a) for a in input("Enter group number 2:"). split()]
 print("Result:")
 setGroup_One= set(group_One)
